feat: Changes implementation to rely fully on emotion and theme-data, separates test concerns
