+++ packages/spacer/package.json
@@ -20,6 +20,7 @@
   ],
   "dependencies": {
     "@hig/themes": "^0.3.0",
+    "@hig/theme-data-poc": "^0.0.9-alpha",
     "classnames": "^2.2.5",
     "prop-types": "^15.6.1"
   },
