+++ packages/spacer/src/Spacer.test.js
@@ -5,29 +5,6 @@ import renderer from "react-test-renderer";
 import Spacer from "./Spacer";
 
 describe("spacer/Spacer", () => {
-  describe("style props", () => {
-    it("allows overriding spacing", () => {
-      const wrapper = mount(<Spacer spacing="m" />);
-      expect(wrapper.find(".hig__spacer-v1").prop("style")).toMatchObject({
-        height: "16px"
-      });
-    });
-
-    it("allows overriding size", () => {
-      const wrapper = mount(<Spacer size="40px" />);
-      expect(wrapper.find(".hig__spacer-v1").prop("style")).toMatchObject({
-        height: "40px"
-      });
-    });
-
-    it("preferences size over spacing", () => {
-      const wrapper = mount(<Spacer spacing="m" size="40px" />);
-      expect(wrapper.find(".hig__spacer-v1").prop("style")).toMatchObject({
-        height: "40px"
-      });
-    });
-  });
-
   describe("snapshot tests", () => {
     const cases = [
       {
