+++ packages/spacer/src/__stories__/getKnobs.js
@@ -1,4 +1,4 @@
-import { text } from "@storybook/addon-knobs/react";
+import { text, select } from "@storybook/addon-knobs/react";
 
 const knobGroupIds = {
   basic: "Basic"
@@ -6,17 +6,30 @@ const knobGroupIds = {
 
 const knobLabels = {
   size: "Size",
-  spacing: "Spacing",
-  display: "Display"
+  spacing: "Spacing"
+};
+
+const spacingOptions = {
+  XXS: "xxs",
+  XS: "xs",
+  S: "s",
+  M: "m",
+  L: "l",
+  XL: "xl",
+  XXL: "xxl"
 };
 
 export default function getKnobs(props) {
-  const { size, spacing, display, ...otherProps } = props;
+  const { size, spacing, ...otherProps } = props;
 
   return {
     ...otherProps,
     size: text(knobLabels.size, size, knobGroupIds.basic),
-    spacing: text(knobLabels.spacing, spacing, knobGroupIds.basic),
-    display: text(knobLabels.display, display, knobGroupIds.basic)
+    spacing: select(
+      knobLabels.spacing,
+      spacingOptions,
+      spacing,
+      knobGroupIds.basic
+    )
   };
 }
