+++ packages/theme-data/src/basics/availableSpacings.js
@@ -1,0 +1,3 @@
+const AVAILABLE_SPACINGS = ["xxs", "xs", "s", "m", "l", "xl", "xxl"];
+
+export default AVAILABLE_SPACINGS;
